requirements.txt
Flask==2.3.3
openai-whisper @ git+https://github.com/openai/whisper.git
python-dotenv==1.0.1
requests==2.31.0
boto3==1.34.103
